+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+- dummy engine valueFormat parameters
+- apiKey checks
+
 ## [1.1.2] - 2025-03-23
 
 ### Changed